Enums to class enums

+++ src/candle/frmgrblconfigurator.cpp
@@ -509,7 +509,7 @@ void frmGrblConfigurator::findParametersToBeSaved(QMap<int, double> settings)
 
     for (QMap<int, double>::iterator it = toBeSaved.begin(); it != toBeSaved.end(); it++) {
         QString command = QString("$%1=%2").arg(it.key()).arg(it.value());
-        qDebug() << m_communicator->sendCommand(CommandSource::System, command);
+        qDebug() << (int) m_communicator->sendCommand(CommandSource::System, command);
         qDebug() << "Sending command" << command;
     }
